82 fix questionscontrollerdestroy

+++ app/controllers/admin/dashboard/questions_controller.rb
@@ -7,9 +7,13 @@ module Admin
 
       def destroy
         question = Question.find(params['id'])
-        question.taggings.destroy_all
-        question.questionnaires.destroy_all
-        question.destroy!
+        if question.children.present?
+          flash[:error] =
+            "Question ##{question.id} has children questions and cannot be deleted. Please remove questions with parent ##{question.id} first and then try again."
+        else
+          question.destroy!
+          flash[:success] = "Successfully deleted question ##{question.id}"
+        end
 
         redirect_to admin_dashboard_questions_path
       end
